[ENHANCEMENT] Switch Clans Without Exiting to Desktop (#3263)
* It works!

would solve #3067 , however I don't know if it's the best, or even good for that matter.

* improved and runs fine

* It works!

would solve #3067 , however I don't know if it's the best, or even good for that matter.

* improved and runs fine

* some minor improvements

clears patrol things and completely empties the events list

* leader den screen fix

also undid the change in utility

* fix frame-long switch clan screen when switching

* add keybind to reload a clan

* update SwitchClanScreen note

* remove reload keybind and other things

* Reload clan without saving

* fix some bugs

* some changes

Mainly just changing how some reused lists are created more than once.

* clan reload is back

It's a debug command this time around, and it also allows the user to save the clan before reloading.

---------

Co-authored-by: J Gynn <48025294+j-gynn@users.noreply.github.com>